+++ operator/CHANGELOG.md
@@ -10,6 +10,10 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 ## Unreleased
 
+### Added
+
+- `spec.adminClient` option to deploy an AIS admin client auto-configured to connect to the cluster.
+  - Supports custom CA trust via `caConfigMap` for TLS-enabled clusters.
 
 ### Removed